backend/requirements.txt
# FastAPI
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0

# Settings
pydantic-settings==2.6.1
email-validator==2.2.0

# Security
python-jose[cryptography]==3.3.0
bcrypt==4.2.1

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1

# Utils
python-dateutil==2.9.0
